requirements.txt
Flask>=3.0.0
PyJWT>=2.7.0
Werkzeug>=3.0.0
